Update create, update and delete blogs to only work for the blog creator.

+++ src/controllers/blogs.controller.js
@@ -39,6 +39,8 @@ async function getDrafts(req, res, next) {
 }
 async function postBlog(req, res, next) {
   try {
+    // Set the creator of the blog from the token payload
+    req.body.author = req.auth.payload.sub;
     const blog = await createBlog(req.body);
     res.send(blog);
   } catch (error) {
@@ -51,13 +53,17 @@ async function patchBlog(req, res, next) {
   const body = req.body;
   const file = req.file;
   try {
-    const update = await cloudUpload(body, file);
     const blog = await fetchBlog(id);
-    if (update.cover || blog.cover) {
-      await cloudDelete(blog.cover);
+    // Only update if the user is the creator of the resource.
+    if (blog && blog.author === req.auth.payload.sub) {
+      const update = await cloudUpload(body, file);
+      if (update.cover || blog.cover) {
+        await cloudDelete(blog.cover);
+      }
+      const updatedBlog = await updateBlog(blog, update);
+      res.send(updatedBlog);
     }
-    const updatedBlog = await updateBlog(blog, update);
-    res.send(updatedBlog);
+    res.sendStatus(404)
   } catch (error) {
     next(error);
   }
@@ -67,7 +73,8 @@ async function deleteBlog(req, res, next) {
   const { id } = req.params;
   try {
     const blog = await fetchBlog(id);
-    if (blog) {
+    // Only delete if the user is the creator of the resource.
+    if (blog && blog.author === req.auth.payload.sub) {
       await cloudDelete(blog.cover);
       await destroyBlog(blog);
       res.sendStatus(200);
